修改流程表之前

+++ src/main/java/com/cjih/learnsystem/answer/AnswerController.java
@@ -88,7 +88,7 @@ public class AnswerController extends BaseController{
 //			answer.setId(UUID.randomUUID().toString());
 			answer.setUserId(userId);
 			answer.setExamId(examId);
-			answer.setQuestionId(q.getId());
+			answer.setQuestionId(q.getId().intValue());
 			answer.setAnswerNo(answer_no);
 			answer.setAnswer("");
 			answer.save();// .insert(answer);
